fix: Fix duplicated code in tests
